+++ lib/src/form_scope.dart
@@ -10,13 +10,13 @@ typedef FastInputDecorator = InputDecoration Function(
     BuildContext context, FastFormField field);
 
 class FastFormScope extends InheritedWidget {
-  FastFormScope({
+  const FastFormScope({
+    Key? key,
     this.adaptive = false,
     required this.builders,
     required Widget child,
     required this.formState,
     required this.inputDecorator,
-    Key? key,
     this.padding,
   }) : super(key: key, child: child);