+++ lib/src/switch/switch.dart
@@ -16,8 +16,6 @@ class FastSwitch extends FastFormField<bool> {
     EdgeInsetsGeometry? contentPadding,
     InputDecoration? decoration,
     bool enabled = true,
-    this.errorBuilder,
-    this.helperBuilder,
     String? helperText,
     required String id,
     bool initialValue = false,
@@ -26,9 +24,11 @@ class FastSwitch extends FastFormField<bool> {
     ValueChanged<bool>? onChanged,
     VoidCallback? onReset,
     FormFieldSetter<bool>? onSaved,
+    FormFieldValidator<bool>? validator,
+    this.errorBuilder,
+    this.helperBuilder,
     this.title,
     this.titleBuilder,
-    FormFieldValidator<bool>? validator,
   }) : super(
           adaptive: adaptive,
           autofocus: autofocus,
@@ -68,7 +68,7 @@ class FastSwitchState extends FastFormFieldState<bool> {
   FastSwitch get widget => super.widget as FastSwitch;
 }
 
-final SwitchTitleBuilder switchTitleBuilder = (FastSwitchState state) {
+Text switchTitleBuilder(FastSwitchState state) {
   return Text(
     state.widget.title!,
     style: TextStyle(
@@ -76,9 +76,9 @@ final SwitchTitleBuilder switchTitleBuilder = (FastSwitchState state) {
       color: state.value! ? Colors.black : Colors.grey,
     ),
   );
-};
+}
 
-final FormFieldBuilder<bool> switchBuilder = (FormFieldState<bool> field) {
+InputDecorator switchBuilder(FormFieldState<bool> field) {
   final state = field as FastSwitchState;
   final widget = state.widget;
   final theme = Theme.of(state.context);
@@ -110,10 +110,9 @@ final FormFieldBuilder<bool> switchBuilder = (FormFieldState<bool> field) {
       );
     }),
   );
-};
+}
 
-final FormFieldBuilder<bool> cupertinoSwitchBuilder =
-    (FormFieldState<bool> field) {
+CupertinoFormRow cupertinoSwitchBuilder(FormFieldState<bool> field) {
   final state = field as FastSwitchState;
   final widget = state.widget;
 
@@ -127,10 +126,9 @@ final FormFieldBuilder<bool> cupertinoSwitchBuilder =
       value: state.value!,
     ),
   );
-};
+}
 
-final FormFieldBuilder<bool> adaptiveSwitchBuilder =
-    (FormFieldState<bool> field) {
+Widget adaptiveSwitchBuilder(FormFieldState<bool> field) {
   final state = field as FastSwitchState;
 
   if (state.adaptive) {
@@ -143,4 +141,4 @@ final FormFieldBuilder<bool> adaptiveSwitchBuilder =
     }
   }
   return switchBuilder(field);
-};
+}