+++ test/dropdown/dropdown_test.dart
@@ -9,7 +9,7 @@ void main() {
     await tester.pumpWidget(getFastTestWidget(
       FastDropdown(
         id: 'dropdown',
-        items: [],
+        items: const [],
       ),
     ));
 
@@ -24,9 +24,9 @@ void main() {
   });
 
   testWidgets('updates FastDropdown', (WidgetTester tester) async {
-    final itemsLength = 3;
+    const testIndex = 2;
+    const itemsLength = 3;
     final items = List.generate(itemsLength, (int index) => 'item $index');
-    final testIndex = 2;
 
     await tester.pumpWidget(getFastTestWidget(
       FastDropdown(
@@ -47,20 +47,21 @@ void main() {
 
     expect(itemsFinder, findsNWidgets(itemsLength * 2));
 
-    await tester.tap(itemsFinder.at(itemsLength + testIndex));
+    await tester.tap(itemsFinder.at(itemsLength + testIndex),
+        warnIfMissed: false);
     await tester.pumpAndSettle();
 
     expect(state.value, items[testIndex]);
   });
 
   testWidgets('validates FastDropdown', (WidgetTester tester) async {
-    final invalidItem = 'invalid item';
-    final errorText = 'Do not touch this';
+    const invalidItem = 'invalid item';
+    const errorText = 'Do not touch this';
 
     await tester.pumpWidget(getFastTestWidget(
       FastDropdown(
         id: 'dropdown',
-        items: ['item', invalidItem],
+        items: const ['item', invalidItem],
         validator: (value) => value == invalidItem ? errorText : null,
       ),
     ));